requirements.txt
alembic==1.1.0
amqp==2.5.1
aniso8601==7.0.0
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
billiard==3.6.1.0
bleach==3.1.0
blinker==1.4
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Django==2.2.4
django-debug-toolbar==2.0
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.10.2
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-paginate==0.5.3
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
greenlet==0.4.15
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10.1
kombu==4.6.4
lxml==4.3.4
Mako==1.1.0
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==7.2.0
mysqlclient==1.4.2.post1
parsel==1.5.1
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pypiwin32==223
python-dateutil==2.8.0
python-editor==1.0.4
python-memcached==1.59
pytz==2019.1
pywin32==224
qiniu==7.2.6
queuelib==1.5.0
redis==3.2.1
requests==2.22.0
Scrapy==1.7.1
selenium==3.141.0
service-identity==18.1.0
shortuuid==0.5.0
six==1.12.0
SQLAlchemy==1.3.8
sqlparse==0.3.0
Twisted==19.2.1
urllib3==1.25.3
vine==1.3.0
virtualenv==16.6.0
virtualenvwrapper-win==1.2.5
w3lib==1.20.0
webencodings==0.5.1
Werkzeug==0.15.5
Whoosh==2.7.4
WTForms==2.2.1
zipp==0.6.0
zope.interface==4.6.0
